Remove-SubtitleSpamLines.ps1
* 12:16 AM 1/28/2023 added banners ; convert func -> script. I've got a lot in verb-io already, and unless it's a
reusable component, there's little bene in sticking it in a module. This is
really a single-purpose deidcated tool == a Script.  convert func -> script. I've got a lot in verb-io already, and unless it's a
reusable component, there's little bene in sticking it in a module. This is
really a single-purpose deidcated tool == a Script.